+++ skoolkit/skool2html.py
@@ -23,11 +23,12 @@ import argparse
 from io import StringIO
 
 from skoolkit import (defaults, SkoolKitError, find_file, show_package_dir,
-                      write, write_line, get_class, normpath, PACKAGE_DIR, VERSION)
+                      write, write_line, get_class, normpath,
+                      PACKAGE_DIR, VERSION, BASE_10, BASE_16)
 from skoolkit.config import get_config, show_config, update_options
 from skoolkit.refparser import RefParser
 from skoolkit.skoolhtml import FileInfo
-from skoolkit.skoolparser import SkoolParser, CASE_UPPER, CASE_LOWER, BASE_10, BASE_16
+from skoolkit.skoolparser import SkoolParser, CASE_UPPER, CASE_LOWER
 
 SEARCH_DIRS = (
     '',